+++ src/main/java/com/example/simple_stock_management/repository/InventoryRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface InventoryRepository extends JpaRepository<Inventory, InventoryKey> {
-    List<Inventory> findByIdItemId(Long itemId);
+    List<Inventory> findByIdItemId(Integer itemId);
 }
