+++ src/main/java/com/example/simple_stock_management/repository/OrderRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OrderRepository extends JpaRepository<Order, Long> {
-    List<OrderResponse> findByItemId(Long itemId);
+public interface OrderRepository extends JpaRepository<Order, Integer> {
+    List<OrderResponse> findByItemId(Integer itemId);
 }
